[Cleanup] Keep InitDeviceProfiler and ReadDeviceProfilerResults public
These APIs are known to be used externally; restore them to
tt_metal_profiler.hpp and leave only the remaining internals in impl/.
