+++ tt_metal/impl/host_api/tt_metal.cpp
@@ -53,7 +53,7 @@
 #include <fmt/format.h>
 #include "llrt.hpp"
 #include <tt-logger/tt-logger.hpp>
-#include "impl/profiler/tt_metal_profiler.hpp"
+#include <tt-metalium/tt_metal_profiler.hpp>
 #include <program.hpp>
 #include "program/program_impl.hpp"
 #include "impl/buffers/semaphore.hpp"
